requirements.txt
bottle==0.12.16
psycopg2==2.8.2
mecab-python3==0.996.2
